+++ config/parameters.js
@@ -9,6 +9,18 @@ const distPath = rootPath + '/dist';
 
 dotenv.config({ path: path.join(rootPath, '.env') });
 
+if (process.env.SOS_API_IDENTIFICATION) {
+	console.warn(`Environment variable SOS_API_IDENTIFICATION found. Will override default credentials from ~/.sosrc`);
+}
+
+if (process.env.SOS_API_SECURITY_TOKEN) {
+	console.warn(`Environment variable SOS_API_SECURITY_TOKEN found. Will override default credentials from ~/.sosrc`);
+}
+
+if (process.env.SOS_ORGANIZATION_UID) {
+	console.warn(`Environment variable SOS_ORGANIZATION_UID found. Will override default credentials from ~/.sosrc`);
+}
+
 module.exports = {
 	environment,
 	name: packageConfig.name,
@@ -21,4 +33,14 @@ module.exports = {
 	},
 	apiUrl: process.env.SOS_API_URL,
 	boxHost: process.env.SOS_BOX_HOST,
+	applet: {
+		uid: process.env.SOS_APPLET_UID,
+		version: process.env.SOS_APPLET_VERSION,
+		name: process.env.SOS_APPLET_NAME,
+	},
+	accountAuth: {
+		tokenId: process.env.SOS_API_IDENTIFICATION,
+		token: process.env.SOS_API_SECURITY_TOKEN,
+	},
+	defaultOrganizationUid: process.env.SOS_ORGANIZATION_UID,
 };