+++ src/Timer/wait.ts
@@ -1,0 +1,4 @@
+
+export default function wait(timeout: number) {
+	return new Promise<void>((resolve: () => void) => setTimeout(() => resolve(), timeout));
+}